remove all the left in the nn.py
