+++ python/paddle/fluid/contrib/tests/test_correlation.py
@@ -77,81 +77,6 @@ def corr(
     return output
 
 
-class TestCorrelationOp(unittest.TestCase):
-    def test_check_output(self):
-        if not fluid.core.is_compiled_with_cuda():
-            return
-        np.random.seed(13)
-        np.set_printoptions(threshold=np.inf)
-        x_shape = (2, 10, 3, 3)
-        x_type = 'float32'
-        x1 = fluid.layers.data(
-            name='x1',
-            shape=x_shape,
-            dtype=x_type,
-            append_batch_size=False,
-            stop_gradient=False,
-        )
-        x2 = fluid.layers.data(
-            name='x2',
-            shape=x_shape,
-            dtype=x_type,
-            append_batch_size=False,
-            stop_gradient=False,
-        )
-
-        x1_np = np.random.randn(2, 3, 4, 5).astype(x_type)
-        x2_np = np.random.randn(2, 3, 4, 5).astype(x_type)
-        out_np = corr(
-            x1_np,
-            x2_np,
-            pad_size=4,
-            kernel_size=1,
-            max_displacement=4,
-            stride1=1,
-            stride2=1,
-        )
-
-        out = fluid.contrib.correlation(
-            x1,
-            x2,
-            pad_size=4,
-            kernel_size=1,
-            max_displacement=4,
-            stride1=1,
-            stride2=1,
-        )
-
-        loss = paddle.mean(out)
-        optimizer = fluid.optimizer.Momentum(0.0001, 0.9)
-        optimizer.minimize(loss)
-
-        place = fluid.CUDAPlace(0)
-        exe = fluid.Executor(place)
-        res = exe.run(
-            feed={'x1': x1_np, 'x2': x2_np}, fetch_list=[out.name, loss.name]
-        )
-
-        np.testing.assert_allclose(res[0], out_np, rtol=1e-05, atol=1e-8)
-
-
-class Net(fluid.dygraph.Layer):
-    def __init__(self, name_scope):
-        super().__init__(name_scope)
-
-    def forward(self, x1, x2):
-        y = fluid.contrib.correlation(
-            x1,
-            x2,
-            pad_size=4,
-            kernel_size=1,
-            max_displacement=4,
-            stride1=1,
-            stride2=1,
-        )
-        return y
-
-
 class TestCorrelationOpDyGraph(unittest.TestCase):
     def test_check_output(self):
         if not fluid.core.is_compiled_with_cuda():
