+++ python/paddle/fluid/tests/unittests/test_dist_fleet_ps12.py
@@ -74,7 +74,7 @@ class TestPSPassWithBow(unittest.TestCase):
         # query
         q = fluid.layers.data(name="1", shape=[1], dtype="int64", lod_level=1)
         # embedding
-        q_emb = fluid.contrib.layers.sparse_embedding(
+        q_emb = paddle.static.nn.sparse_embedding(
             input=q,
             size=[dict_dim, emb_dim],
             param_attr=fluid.ParamAttr(
@@ -102,7 +102,7 @@ class TestPSPassWithBow(unittest.TestCase):
         # pt
         pt = fluid.layers.data(name="2", shape=[1], dtype="int64", lod_level=1)
         # embedding
-        pt_emb = fluid.contrib.layers.sparse_embedding(
+        pt_emb = paddle.static.nn.sparse_embedding(
             input=pt,
             size=[dict_dim, emb_dim],
             param_attr=fluid.ParamAttr(
@@ -129,7 +129,7 @@ class TestPSPassWithBow(unittest.TestCase):
         # nt
         nt = fluid.layers.data(name="3", shape=[1], dtype="int64", lod_level=1)
         # embedding
-        nt_emb = fluid.contrib.layers.sparse_embedding(
+        nt_emb = paddle.static.nn.sparse_embedding(
             input=nt,
             size=[dict_dim, emb_dim],
             param_attr=fluid.ParamAttr(
