+++ python/paddle/fluid/tests/unittests/test_dist_fleet_ps13.py
@@ -77,7 +77,7 @@ class TestPSPassWithBow(unittest.TestCase):
             name="query_ids", shape=[1], dtype="int64", lod_level=1
         )
         # embedding
-        q_emb = fluid.contrib.layers.sparse_embedding(
+        q_emb = paddle.static.nn.sparse_embedding(
             input=q,
             size=[dict_dim, emb_dim],
             param_attr=fluid.ParamAttr(
@@ -107,7 +107,7 @@ class TestPSPassWithBow(unittest.TestCase):
             name="pos_title_ids", shape=[1], dtype="int64", lod_level=1
         )
         # embedding
-        pt_emb = fluid.contrib.layers.sparse_embedding(
+        pt_emb = paddle.static.nn.sparse_embedding(
             input=pt,
             size=[dict_dim, emb_dim],
             param_attr=fluid.ParamAttr(
@@ -136,7 +136,7 @@ class TestPSPassWithBow(unittest.TestCase):
             name="neg_title_ids", shape=[1], dtype="int64", lod_level=1
         )
         # embedding
-        nt_emb = fluid.contrib.layers.sparse_embedding(
+        nt_emb = paddle.static.nn.sparse_embedding(
             input=nt,
             size=[dict_dim, emb_dim],
             param_attr=fluid.ParamAttr(
