+++ python/paddle/fluid/tests/unittests/test_layers.py
@@ -2156,55 +2156,6 @@ class TestBook(LayerTest):
             )
             return out
 
-    def test_partial_sum(self):
-        with self.static_graph():
-            x = fluid.data(name="x", shape=[None, 3], dtype="float32")
-            y = fluid.data(name="y", shape=[None, 3], dtype="float32")
-            sum = fluid.contrib.layers.partial_sum(
-                [x, y], start_index=0, length=2
-            )
-            return sum
-
-    def test_batch_fc(self):
-        with self.static_graph():
-            input = fluid.data(name="input", shape=[16, 2, 3], dtype="float32")
-            out = fluid.contrib.layers.batch_fc(
-                input=input,
-                param_size=[16, 3, 10],
-                param_attr=fluid.ParamAttr(
-                    learning_rate=1.0,
-                    name="w_0",
-                    initializer=fluid.initializer.Xavier(uniform=False),
-                ),
-                bias_size=[16, 10],
-                bias_attr=fluid.ParamAttr(
-                    learning_rate=1.0,
-                    name="b_0",
-                    initializer=fluid.initializer.Xavier(uniform=False),
-                ),
-                act="relu",
-            )
-        return out
-
-    def test_rank_attention(self):
-        with self.static_graph():
-            input = fluid.data(name="input", shape=[None, 2], dtype="float32")
-            rank_offset = fluid.data(
-                name="rank_offset", shape=[None, 7], dtype="int32"
-            )
-            out = fluid.contrib.layers.rank_attention(
-                input=input,
-                rank_offset=rank_offset,
-                rank_param_shape=[18, 3],
-                rank_param_attr=fluid.ParamAttr(
-                    learning_rate=1.0,
-                    name="ubm_rank_param.w_0",
-                    initializer=fluid.initializer.Xavier(uniform=False),
-                ),
-                max_rank=3,
-            )
-            return out
-
     def test_sequence_enumerate(self):
         # TODO(minqiyang): dygraph do not support lod now
         with self.static_graph():
