+++ python/paddle/static/nn/__init__.py
@@ -36,10 +36,10 @@ from ...tensor.creation import create_parameter  # noqa: F401
 from .loss import nce  # noqa: F401
 from .common import prelu  # noqa: F401
 from .common import layer_norm  # noqa: F401
+from .common import sparse_embedding  # noqa: F401
 
 
 from ...fluid.input import embedding  # noqa: F401
-from ...fluid.contrib.layers import sparse_embedding  # noqa: F401
 from ...fluid.layers import StaticRNN  # noqa: F401
 
 from ...fluid.layers.sequence_lod import sequence_conv  # noqa: F401
